requirements.txt
anyio==3.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.5.3
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.10.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
dj-stripe==2.5.1
Django==3.1.7
django-countries==7.2.1
django-debug-toolbar==1.10.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Js2Py==0.71
json5==0.9.5
jsonfield==3.1.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
opencv-python==4.5.4.60
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pipwin==0.5.1
platformdirs==2.4.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
PyAudio @ file:///C:/Users/knira/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
Pygments==2.9.0
pyjsparser==2.7.1
PyJWT==2.3.0
pylint==2.7.4
pyparsing==2.4.7
PyPrind==2.11.3
pyrsistent==0.17.3
pySmartDL==1.3.4
python-dateutil==2.8.1
python-decouple==3.1
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pywin32==301
pywinpty==1.1.2
pyzmq==22.1.0
requests==2.25.1
scipy==1.7.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
SQLAlchemy==1.4.27
sqlparse==0.4.1
stripe==2.64.0
terminado==0.10.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.5
virtualenv==20.4.3
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.2
wrapt==1.12.1
